gui: improve settings, resizing and update Pugl
These 3 changes unfortunately come as one.
- Use settings descriptor objects for the GUI settings and move this
  into a separate file.
- Resizing is greatly improved and the related code refactored. We now
  correctly disassociate the actual window size and the desired window
  size from the settings. We also properly handle the various situations
  where the window sizes changes or need to be changed, etc.
- Update pugl to make use of the clearer API for setting size and
  position.

+++ src/plugin/gui/gui2_common_modal_panel.hpp
@@ -246,8 +246,8 @@ PUBLIC bool TextButton(GuiBoxSystem& builder, Box parent, String text, String to
     return button.button_fired;
 }
 
-PUBLIC Optional<int>
-IntField(GuiBoxSystem& builder, Box parent, String label, f32 width, int value, int min, int max) {
+PUBLIC Optional<s64>
+IntField(GuiBoxSystem& builder, Box parent, String label, f32 width, s64 value, s64 min, s64 max) {
     bool changed = false;
     auto const container = DoBox(builder,
                                  {
@@ -308,7 +308,7 @@ IntField(GuiBoxSystem& builder, Box parent, String label, f32 width, int value, 
         if (text_input.text_input_result) {
             auto const new_value = ParseInt(text_input.text_input_result->text, ParseIntBase::Decimal);
             if (new_value.HasValue()) {
-                value = (int)Clamp<s64>(new_value.Value(), min, max);
+                value = Clamp<s64>(new_value.Value(), min, max);
                 changed = true;
             }
         }
